+++ frontend/index.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html lang="es">
+<head>
+    <meta charset="UTF-8">
+    <title>Gestión de Riegos</title>
+    <link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.4/dist/leaflet.css" />
+    <style>
+        #map { height: 300px; margin-bottom: 1rem; }
+        table { border-collapse: collapse; width: 100%; }
+        th, td { border: 1px solid #ccc; padding: 4px; }
+    </style>
+</head>
+<body>
+    <h1>Gestión de Riegos</h1>
+    <div id="map"></div>
+    <div id="riego-tab">
+        <h2>Riego</h2>
+        <label>Lote seleccionado: <span id="lote-display">Ninguno</span></label><br>
+        <label>Fecha: <input type="date" id="fecha"></label>
+        <button id="guardar">Guardar Riego</button>
+        <h3>Historial</h3>
+        <table id="tabla-riegos">
+            <thead><tr><th>ID</th><th>Lote</th><th>Fecha</th><th>Nota</th></tr></thead>
+            <tbody></tbody>
+        </table>
+    </div>
+    <script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js"></script>
+    <script src="app.js"></script>
+</body>
+</html>
